Implement S3 snapshots-specific retry policy default
Also remove the skip-if-default check which appears to not do much now that we have a default retry
policy.

+++ crates/object-store-util/src/lib.rs
@@ -212,7 +212,7 @@ fn from_retry_policy(retry_policy: &RetryPolicy) -> RetryConfig {
                 max_backoff: max_interval
                     .unwrap_or_else(|| Duration::from_secs(15).into())
                     .into(),
-                base: f64::from(f32::from(*factor)),
+                base: f64::from(*factor),
             },
             retry_timeout: max_interval
                 .map(|interval| interval.into())
